+++ src/index.js
@@ -22,8 +22,14 @@ module.exports = function (options) {
       rootPath + '/config/defaults.json'
     ]);
 
+  // Util service
+  base.utils = options.utils || require('./modules/utils')(base);
+
   // Logger service
   base.logger = options.logger || require('./modules/logger')(base);
+
+  // Cache service
+  base.cache = options.cache || require('./modules/cache')(base);
 
   // Events service
   base.events = options.events || require('./modules/events')(base);