+++ src/modules/services/index.js
@@ -8,6 +8,12 @@ const sessionCache = require('session-cache');
 
 module.exports = function (base) {
 
+  const service = {
+    name: base.config.get('services:name'),
+    version: base.config.get('services:version'),
+    operations: new Set()
+  };
+
   const wreck = Wreck.defaults({
     headers: {
       'Content-Type': 'application/json',
@@ -18,8 +24,17 @@ module.exports = function (base) {
   const gatewayHost = base.config.get('gateway:host');
   const gatewayPort = base.config.get('gateway:port');
   const gatewayBasePath = base.config.get('gateway:path');
+  const gatewayUrlOverrride = base.config.get('gateway:gatewayUrlOverrride');
+  const remoteCallsTimeout = base.config.get('gateway:timeout');
+  let getGatewayBaseUrl;
+  if (gatewayUrlOverrride) {
+    getGatewayBaseUrl = base.utils.loadModule('gateway:gatewayUrlOverrride');
+  } else {
+    const gatewayBaseUrl = `http://${gatewayHost}:${gatewayPort}`;
+    getGatewayBaseUrl = () => gatewayBaseUrl;
+  }
+
   const serviceBasePath = base.config.get('services:path');
-  const gatewayBaseUrl = `http://${gatewayHost}:${gatewayPort}`;
 
   const getOperationUrl = (basePath, serviceName, serviceVersion, operationName, operationPath) =>
     `${basePath}/${serviceName}/${serviceVersion}${operationPath !== undefined ? operationPath : '/' + operationName}`;
@@ -41,16 +56,11 @@ module.exports = function (base) {
     return { serviceName, serviceVersion, operationName };
   };
 
-  const service = {
-    name: base.config.get('services:name'),
-    version: base.config.get('services:version'),
-    operations: new Set()
-  };
-
   const server = service.server = new Hapi.Server();
   server.connection({
     host: base.config.get('services:host'),
-    port: base.config.get('services:port')
+    port: base.config.get('services:port'),
+    routes: { cors: true }
   });
 
   // Custom error responses
@@ -143,6 +153,7 @@ module.exports = function (base) {
       'x-request-id': session.get('x-request-id'),
       authorization: session.get('authorization')
     };
+    Object.assign(headers, config.headers);
 
     let {serviceName, serviceVersion, operationName} = splitOperationName(config.name);
     const operationFullName = getOperationFullName(serviceName, serviceVersion, operationName);
@@ -157,6 +168,7 @@ module.exports = function (base) {
         method: config.method || 'POST',
         headers: headers
       }).then(response => {
+        //return Promise.resolve(response.result, response);
         return new Promise(resolve => {
           return resolve(response.result, response);
         });
@@ -164,7 +176,7 @@ module.exports = function (base) {
     } else {
       // It's a remote operation
       return new Promise((resolve, reject) => {
-        const operationUrl = getOperationUrl(gatewayBaseUrl + gatewayBasePath, serviceName, serviceVersion, operationName, config.path);
+        const operationUrl = getOperationUrl(getGatewayBaseUrl(serviceName, serviceVersion, operationName) + gatewayBasePath, serviceName, serviceVersion, operationName, config.path);
         if (base.logger.isDebugEnabled()) base.logger.debug(`[services] calling [${operationMethod}] ${operationUrl} with ${JSON.stringify(msg)}`);
         wreck.request(
           operationMethod,
@@ -172,9 +184,10 @@ module.exports = function (base) {
           {
             payload: JSON.stringify(msg),
             headers: headers,
-            timeout: config.timeout || 1000
+            timeout: config.timeout || remoteCallsTimeout
           },
           (error, response) => {
+            if (error) return reject(error);
             Wreck.read(response, { json: 'smart' }, (error, payload) => {
               if (error) return reject(error);
               return resolve(payload, response);
@@ -198,21 +211,75 @@ module.exports = function (base) {
     }
   };
 
-  // Routes handler
-  const routeHandler = (handler) => (request, reply) => {
-    // Mix body payload/params/query
-    let payload = request.payload || {};
-    Object.assign(payload, request.params);
-    Object.assign(payload, request.query);
-    // Create CID
-    if (!request.headers['x-request-id']) {
-      request.headers['x-request-id'] = uuid();
+  // Cache responses
+  server.ext('onPreResponse', (request, reply) => {
+    const response = request.response;
+    if (response.isBoom || !request.headers['mb-cache']) {
+      return reply.continue();
     }
-    // Store CID & Authorization token in session
-    session.set('x-request-id', request.headers['x-request-id']);
-    session.set('authorization', request.headers['authorization']);
+    const cache = base.cache.get(request.headers['mb-cache']);
 
-    return handler.call(this, payload, reply, request);
+    cache.set(request.headers['mb-cache-key'], {
+      statusCode: response.statusCode || 200,
+      payload: response.source
+    });
+    return reply.continue();
+  });
+
+  // Routes handler
+  const routeHandler = (cacheOptions, handler) => {
+    return (request, reply) => {
+      // Mix body payload/params/query
+      let payload = request.payload || {};
+      Object.assign(payload, request.params);
+      Object.assign(payload, request.query);
+      // Create CID
+      if (!request.headers['x-request-id']) {
+        request.headers['x-request-id'] = uuid();
+      }
+      // Store CID & Authorization token in session
+      session.set('x-request-id', request.headers['x-request-id']);
+      session.set('authorization', request.headers['authorization']);
+
+      // Cache the results id configured
+      if (cacheOptions) {
+        const key = (cacheOptions.keyGenerator ? (cacheOptions.keyGenerator(payload) + ':') : '') + base.utils.hash(payload);
+
+        // Verify the no-cache header to bypass the cache
+        let noStore = false;
+        if (request.headers['cache-control']) {
+          noStore = Wreck.parseCacheControl(request.headers['cache-control'])['no-store'];
+        }
+        if (noStore) {
+          // Set the headers to store the result
+          request.headers['mb-cache'] = cacheOptions.name;
+          request.headers['mb-cache-key'] = key;
+          // Execute the operation
+          return handler.call(this, payload, reply, request);
+        }
+
+        // Try to get the result from cache
+        const cache = base.cache.get(cacheOptions.name);
+        cache.get(key)
+          .then((value) => {
+            if (value) {
+              // If the result was on the cache, just return it.
+              return reply(value.payload).code(value.statusCode || 200);
+            }
+            // The result was not in the cache, set the headers to store the result
+            request.headers['mb-cache'] = cacheOptions.name;
+            request.headers['mb-cache-key'] = key;
+            // Execute the operation
+            return handler.call(this, payload, reply, request);
+          })
+          .catch(error => {
+            return reply(err);
+          });
+      } else {
+        // The operation is not cacheable
+        return handler.call(this, payload, reply, request);
+      }
+    };
   };
 
   // Routes style
@@ -235,11 +302,16 @@ module.exports = function (base) {
     base.logger.info(`[services] added service [${operationFullName}] in [${operationMethod}][${operationUrl}]`);
     // Add the operation to this service operations
     service.operations.add(operationFullName);
+    // Create cache
+    if (op.cache) {
+      op.cache.name = op.cache.name || operationFullName;
+      base.cache.create(op.cache.name, op.cache.options);
+    }
     // Add the Hapi route, mixing parameters and payload to call the handler
     server.route({
       method: operationMethod,
       path: operationUrl,
-      handler: routeHandler(op.handler),
+      handler: routeHandler(op.cache, op.handler),
       config: op.config || routeConfig(op.schema, op.scope || defaultScope)
     });
   };
@@ -260,24 +332,6 @@ module.exports = function (base) {
       return reply({ answer: 'pong' });
     }
   })
-
-  // Load a module
-  service.loadModule = function (key) {
-    if (base.logger.isDebugEnabled()) base.logger.debug(`[services] loading module from ${key}`);
-    if (!key) return null;
-    const name = base.config.get(key);
-    if (name.startsWith('.')) {
-      const modulePath = `${base.config.get('rootPath')}/${name}`;
-      try {
-        return require(modulePath)(base)
-      } catch (e) {
-        base.logger.error(`[services] module '${modulePath}' not found`)
-        return false;
-      }
-    } else {
-      return require(name)(base);
-    }
-  };
 
   // session-cache to store authorization and CID
   var session = sessionCache('microbase');